backend/requirements.txt
fastapi
uvicorn[standard]
pymupdf
langchain
openai
pinecone
httpx
python-multipart
pydantic
python-dotenv
